Docs invoice attributes
