+++ models/invoice/HasCalculatedAttributes.php
@@ -1,0 +1,9 @@
+<?php namespace Responsiv\Pay\Models\Invoice;
+
+/**
+ * HasCalculatedAttributes
+ */
+trait HasCalculatedAttributes
+{
+
+}
